requirements.txt
tabulate==0.9.0
numpy==1.24.2
pandas==1.5.3
python-dateutil==2.8.2
pytz==2022.7.1
six==1.16.0
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.11.2
bleach==6.0.0
certifi==2022.12.7
charset-normalizer==2.1.1
contourpy==1.0.7
cycler==0.11.0
dataframe_image==0.1.5
defusedxml==0.7.1
fastjsonschema==2.16.2
fonttools==4.38.0
frozenlist==1.3.3
idna==3.4
jinja2==3.1.2
jsonschema==4.17.3
jupyter-client==8.0.2
jupyter-core==5.2.0
jupyterlab-pygments==0.2.2
kiwisolver==1.4.4
markupsafe==2.1.2
matplotlib==3.6.3
mistune==2.0.4
multidict==6.0.4
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
packaging==23.0
pandocfilters==1.5.0
pillow==9.4.0
platformdirs==2.6.2
pygments==2.14.0
pyparsing==3.0.9
pyrsistent==0.19.3
pywin32==305
pyzmq==25.0.0
requests==2.28.2
soupsieve==2.3.2.post1
tinycss2==1.2.1
tornado==6.2
traitlets==5.9.0
urllib3==1.26.14
webencodings==0.5.1
yarl==1.8.2